pod lib lint 警告修复

+++ SFRouter/Classes/SFModels/SFRouterModels.m
@@ -26,22 +26,22 @@ NSString * const SFRouterDomain = @"com.shawnfly.router";
 - (NSError *)checkEqualForInfoItem:(SFRouterInfoItem *)infoItem {
     NSError *error = nil;
     if (infoItem.isAction != self.isAction) {
-        NSString *errMsg = [NSString stringWithFormat:@"路由类型对应不上，errPath:%@  %ld行", self.filePath, self.lineNum];
+        NSString *errMsg = [NSString stringWithFormat:@"路由类型对应不上，errPath:%@  %ld行", self.filePath, (long)self.lineNum];
         error = [NSError errorWithDomain:SFRouterDomain code:-1 userInfo:@{
             NSLocalizedFailureReasonErrorKey : errMsg
         }];
     } else if (![infoItem.returnTypeName isEqualToString:self.returnTypeName]) {
-        NSString *errMsg = [NSString stringWithFormat:@"路由返回值类型对应不上，errPath:%@  %ld行", self.filePath, self.lineNum];
+        NSString *errMsg = [NSString stringWithFormat:@"路由返回值类型对应不上，errPath:%@  %ld行", self.filePath, (long)self.lineNum];
         error = [NSError errorWithDomain:SFRouterDomain code:-1 userInfo:@{
             NSLocalizedFailureReasonErrorKey : errMsg
         }];
     } else if (infoItem.paramsCount != self.paramsCount) {
-        NSString *errMsg = [NSString stringWithFormat:@"路由参数个数对应不上，errPath:%@  %ld行", self.filePath, self.lineNum];
+        NSString *errMsg = [NSString stringWithFormat:@"路由参数个数对应不上，errPath:%@  %ld行", self.filePath, (long)self.lineNum];
         error = [NSError errorWithDomain:SFRouterDomain code:-1 userInfo:@{
             NSLocalizedFailureReasonErrorKey : errMsg
         }];
     } else if (![infoItem.selName isEqualToString:self.selName]) {
-        NSString *errMsg = [NSString stringWithFormat:@"路由方法名对应不上，errPath:%@  %ld行", self.filePath, self.lineNum];
+        NSString *errMsg = [NSString stringWithFormat:@"路由方法名对应不上，errPath:%@  %ld行", self.filePath, (long)self.lineNum];
         error = [NSError errorWithDomain:SFRouterDomain code:-1 userInfo:@{
             NSLocalizedFailureReasonErrorKey : errMsg
         }];
@@ -50,7 +50,7 @@ NSString * const SFRouterDomain = @"com.shawnfly.router";
             BOOL result = [infoItem.params[index] checkEqualForParamItem:self.params[index]];
             if (!result) {
                 // 添加路径信息
-                NSString *errMsg = [NSString stringWithFormat:@"路由参数校验失败，errPath:%@  %ld行", self.filePath, self.lineNum];
+                NSString *errMsg = [NSString stringWithFormat:@"路由参数校验失败，errPath:%@  %ld行", self.filePath, (long)self.lineNum];
                 error = [NSError errorWithDomain:SFRouterDomain code:-1 userInfo:@{
                     NSLocalizedFailureReasonErrorKey : errMsg
                 }];
